Remove dependency to Freemail and WebOfTrust
