--- ceno-freenet/src/org/freenetproject/freemail/utils/SimpleFieldSetFactory.java
@@ -1,36 +1,0 @@
-/*
- * SimpleFieldSetFactory.java
- * This file is part of Freemail
- * Copyright (C) 2011 Martin Nyhus
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-
-package org.freenetproject.freemail.utils;
-
-import freenet.support.SimpleFieldSet;
-
-public class SimpleFieldSetFactory {
-	private final SimpleFieldSet sfs = new SimpleFieldSet(false);
-
-	public SimpleFieldSetFactory put(String field, String value) {
-		sfs.putOverwrite(field, value);
-		return this;
-	}
-
-	public SimpleFieldSet create() {
-		return sfs;
-	}
-}
